eventsub/pyproject.toml
[tool.poetry]
name = "felpsbot-eventsub"
version = "0.1.0"
description = "Service for handling Twitch EventSub events"
authors = ["Pármenas Haniel <parmenashp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
asyncpg = "^0.26.0"
httpx = "^0.23.0"
loguru = "^0.6.0"
redis = "^4.3.4"
Authlib = "^1.1.0"
orjson = "^3.8.0"
iso8601 = "^1.1.0"
uvicorn = { extras = ["standard"], version = "^0.18.3" }

[tool.poetry.dev-dependencies]
black = "^22.10.0"
python-dotenv = "^0.21.0"

[tool.black]
line-length = 119

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
